requirements.txt
streamlit==1.37.0
Pillow==10.4.0
numpy==1.26.4
scipy==1.11.4
python-dotenv==1.0.1
requests==2.32.3
fastapi==0.112.0
uvicorn==0.30.5
python-multipart==0.0.9
tenacity==8.5.0
edge-tts==6.1.12
moviepy==1.0.3
langgraph==0.2.34
langchain-core==0.3.7
openai==1.43.0
pydantic==2.9.2
tiktoken==0.7.0


